Use rm constraint instead of g for movd arg
When using the -flto=auto flag, gcc is for some reason unable to resolve
the g asm constraint, resulting in a link error. movd args used in
ualgobase are fill algo arguments and can be fully constrained by rm.
The fill value is a function parameter, which is a register on x86_64
and a stack address on i386.