Remove the ability to use with libstdc++
libstdc++ and uSTL follow very different programming philosophies,
and mixing them often produces a variety of unexpected behaviors.
It is important to emphasize that uSTL replaces libstdc++ rather than
adds to it, requiring not just changing which header file you include,
but possibly also the entire design of your application and the thought
process that you used to come up with it.