Replace --without-bounds with --with-bstrxept
--without-bounds was disabling runtime checks on binary streams istream
and ostream. When these streams are used properly, by measuring and
preallocating the output space, no runtime checking is needed.
istringstream and ostringstream are not affected by this option.
The new option name is intended to clarify that binary stream runtime
checks are not the recommended behavior and are off by default.