Remove simd optimized copy and fill
Back in 2005 libc memcpy was not optimized this way, now it is, so there
is no reason to reimplement it here. Optimizing copy is specific to each
CPU these days, so the approach used by uSTL was likely not the fastest
any more anyway. The only asm I'm keeping here is stosw and stosl
optimizations for fill.