Stop looking for libsupc++; just use it
libsupc++ is frequently well hidden and not finding it has been a
recurring source of build failures. Because it should always be present
it is better to let gcc report its absence instead. The only remaining
configure code for this is to replace with libc++abi.a with building
with clang on macs.